Time: 0 ms (100.00%) | Memory: 8.5 MB (97.16%) - LeetSync

+++ 29-divide-two-integers/divide-two-integers.cpp
@@ -7,8 +7,8 @@ public:
         int sign=1;
         if(dividend<0 && divisor>0)sign=-1;
         if(dividend>0 && divisor<0)sign=-1;
-        long long n = llabs((long long)dividend);
-long long d = llabs((long long)divisor);
+        long long n=llabs(dividend);
+        long long d=llabs(divisor);
         long long ans=0;
         while(n>=d){
             int cnt=0;
